Finaly fixes on pull request
