+++ src/main/java/org/fogbowcloud/manager/core/threads/SpawningMonitor.java
@@ -9,7 +9,6 @@ import org.fogbowcloud.manager.core.constants.DefaultConfigurationConstants;
 import org.fogbowcloud.manager.core.datastructures.OrderStateTransitioner;
 import org.fogbowcloud.manager.core.datastructures.SharedOrderHolders;
 import org.fogbowcloud.manager.core.exceptions.OrderStateTransitionException;
-import org.fogbowcloud.manager.core.instanceprovider.InstanceProvider;
 import org.fogbowcloud.manager.core.models.linkedList.SynchronizedDoublyLinkedList;
 import org.fogbowcloud.manager.core.models.orders.Order;
 import org.fogbowcloud.manager.core.models.orders.OrderState;
@@ -24,16 +23,12 @@ public class SpawningMonitor extends Thread {
 
 	private static final Logger LOGGER = Logger.getLogger(SpawningMonitor.class);
 
-	private InstanceProvider localInstanceProvider;
 	private SynchronizedDoublyLinkedList spawningOrderList;
-
 	private TunnelingServiceUtil tunnelingService;
 	private SshConnectivityUtil sshConnectivity;
-
 	private Long sleepTime;
 
-	public SpawningMonitor(InstanceProvider localInstanceProvider, Properties properties) {
-		this.localInstanceProvider = localInstanceProvider;
+	public SpawningMonitor(Properties properties) {
 		SharedOrderHolders sharedOrderHolders = SharedOrderHolders.getInstance();
 		this.spawningOrderList = sharedOrderHolders.getSpawningOrdersList();
 
@@ -57,9 +52,8 @@ public class SpawningMonitor extends Thread {
 				if (order != null) {
 					this.processSpawningOrder(order);
 				} else {
-					spawningOrderList.resetPointer();
-					LOGGER.info("There is no spawning order to be processed, sleeping for " + this.sleepTime
-							+ " milliseconds");
+					LOGGER.info("There is no spawning order to be processed, sleeping for " + this.sleepTime + " milliseconds");
+					this.spawningOrderList.resetPointer();
 					Thread.sleep(this.sleepTime);
 				}
 			} catch (Throwable e) {
@@ -76,14 +70,11 @@ public class SpawningMonitor extends Thread {
 		synchronized (order) {
 			OrderState orderState = order.getOrderState();
 			if (orderState.equals(OrderState.SPAWNING)) {
-				LOGGER.info("Trying to get an instance for order [" + order.getId() + "]");
+				LOGGER.info("Trying to process an instance for order [" + order.getId() + "]");
 				try {
-					
 					this.processInstance(order);
-					
 				} catch (Exception e) {
-					LOGGER.error("Error while trying to get an instance for order: " + System.lineSeparator() + order,
-							e);
+					LOGGER.error("Error while trying to process an instance for order: " + System.lineSeparator() + order, e);
 				}
 			} else {
 				LOGGER.info("This order state is not spawning for order [" + order.getId() + "]");
@@ -92,28 +83,27 @@ public class SpawningMonitor extends Thread {
 	}
 
 	private void processInstance(Order order) throws OrderStateTransitionException {
-		// This method does not synchronize the order object because it is private and
-		// can only be called by the processSpawningOrder method.
-
-		OrderInstance orderInstance = order.getOrderInstance();
+		OrderInstance orderInstance = order.getOrderInstance();		
 		
 		if (order.getType().equals(OrderType.COMPUTE)) {
 			if (orderInstance.getState().equals(InstanceState.FAILED)) {
-				// TODO: log
+				LOGGER.info("The compute instance state is failed for order [" + order.getId() + "]");
 				OrderStateTransitioner.transition(order, OrderState.FAILED);
+			
 			} else if (orderInstance.getState().equals(InstanceState.ACTIVE)) {
 				LOGGER.info("Processing active compute instance for order [" + order.getId() + "]");
-
+				
 				ComputeOrderInstance computeOrderInstance = (ComputeOrderInstance) orderInstance;
-
 				this.setTunnelingServiceAddresses(order, computeOrderInstance);
-				if (isActiveConnectionFromInstance(computeOrderInstance)) {
-					OrderStateTransitioner.transition(order, OrderState.FULFILLED);
+				
+				if (this.isActiveConnectionFromInstance(computeOrderInstance)) {
+					OrderStateTransitioner.transition(order, OrderState.FULFILLED);				
 				} else {
 					LOGGER.warn("Failed attempt to communicate with ssh connectivity for order [" + order.getId() + "]");
 				}
+				
 			} else {
-				LOGGER.info("The compute instance is inactive for order [" + order.getId() + "]");
+				LOGGER.info("The compute instance state is inactive for order [" + order.getId() + "]");
 			}
 		}
 	}
